pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "screenshot"
version = "1.0.0"
description = "Take screenshots on macOS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"pyobjc-framework-ScreenCaptureKit",
	"rich",
]

[[project.authors]]
name = "Alex DeLorenzo"

[project.license]
file = "LICENSE"

[project.urls]
Homepage = "https://github.com/thismachinechills/pyscreencapture"

[project.scripts]
screenshot = "screenshot.screencapture:main"

[tool.setuptools.packages.find]
include = ["screenshot*"]
